_src/* - port content into markdown (plus HTML)
Strip out all the boilerplate HTML stuff which is now in the layouts.

Use a "pagination" hack to duplicate the index.html page on
co/index.html

Leave enough HTML in the markdown to retain the current look (more or
less - the paragraph spacing on the index has increased because I
deleted one level of <div> wrapping, but I actually think this is
eaier to read so I'm leaving it.