_src/index.md - absolutise the link to the registration form
Currently, our duplicated index.html under op/ (which aims to allow
the pauper's .coop link, https://sepheo.co/op to work) no longer links
to the registration form because of it's different location.

Using an absolute URL fixes that.

Also, as we earlier enabled the HTML Base plugin, using an absolute
UTL will now not prevent the site being hosted on a subdirectory (as
might or might not be needed, we aim to leave the option open as a
good practice).