.github/workflows/ - alter workflow to create previews for PRs
And clean up after PRs closed.

Adapted from https://daiyi.co/blog/pr-previews-for-github-pages/

Needed to adapt for eleventy and it's --pathprefix option. Made some
reduction of the number of environment variables also.

The `--pathprefix` option ensures that absolute links within the site
get adjusted correctly for the deployment, courtesy of the
[EleventyHtmlBasePlugin][1] (External and relative links are not
affected.)

Note, the `--pathprefix=...` option *must* have an equals delimiter,
and the prefix will be inserted verbatim. Whereas the
`destination_dir` option for `peaceiris/actions-gh-pages@4` insists
that the path is relative. Therefore the `PR_PATH_PREFIX` variable is
set to a relative path, and a leading slash delimiter is added
elsewhere when required.

[1]: https://www.11ty.dev/docs/plugins/html-base/