_src/membership.adoc - don't mention "affilate" in context of registration
If they've registered already, no need to again, of course.

But if they've not registered already, for the question about
aspiration they should probably say "member", not as an
"affiliate". So adjust the wording not to imply the latter.

(If they've registered already, no current way to change the membership
on the registration form except an admin editing it.)