.github/{actions,workflows} - deploy branch previews as well as PR previews
Problem:
- we want to do A/B testing of website designs
- PR previews kind of work for this
- But sometimes we want to preview more than just purely PRs - e.g. previous versions

Solution:
- Also preview branches.
- To avoid previewing branches which are not intended as previews,
  require a prefix `preview/`

Implementing this with the existing workflow files leads to complexity
and unreadability, because they have a lot of conditions.

So this version defines a composite action for the common build steps,
and three workflows for building and deploying:
- deploy-main-to-ghpages: just for deploying the `main` branch
- deploy-pr-to-ghpages: just for deploying PRs
- deploy-previews-to-ghpages: just for deploying preview/ branchces

And two for erasing old previews:
- erase-pr-preview: erases a PR preview on PR closure.
- erase-branch-preview: erases a branch preview on branch deletion

We also fix a problem in the previous workflows, in which a deploy of
the main branch would remove everything in the gh-pages branch and
replace with the new site. This means any ongoing PRs get removed. A
consequence is that when any PR is merged, all the other PRs will be
expunged. Instead, by using an alternative deploy action
implementation (JamesIves/github-pages-deploy-action) which allows the
cleaning of the target directory to have exceptions. We set the build
to exclude the `_preview` directory from cleaning, and thereby
preserve any ongoing previews.